card

+++ app/Http/Controllers/BasketController.php
@@ -30,8 +30,16 @@ class BasketController extends Controller
         } else {
             $order = Order::find($orderId);
         }
-        $order->products()->attach($productId);
 
+        if ($order->products->contains($productId))
+        {
+            $pivotRow = $order->products()->where('product_id', $productId)->first()->pivot;
+            $pivotRow->count++;
+            $pivotRow->update();
+        } else {
+            $order->products()->attach($productId);
+        }
+        
         return redirect()->route('basket');
     }
     public function basketRemove($productId)
@@ -42,7 +50,19 @@ class BasketController extends Controller
             return redirect()->route('basket');
         }
         $order = Order::find($orderId);
-        $order->products()->detach($productId);
+
+        if ($order->products->contains($productId))
+        {
+            $pivotRow = $order->products()->where('product_id', $productId)->first()->pivot;
+            if ($pivotRow->count < 2)
+            {
+                $order->products()->detach($productId);
+            } else {
+                $pivotRow->count--; 
+                $pivotRow->update();
+            }
+        } 
+
         return redirect()->route('basket');
     }
 }
